dom/payment/tests/mochitest/test_mozpay_callbacks.html

<!DOCTYPE HTML>
<html>
<!--
  https://bugzilla.mozilla.org/show_bug.cgi?id=1097928
-->
<html>
<head>
  <title>Test for navigator.mozPay. Bug 1097928</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript;version=1.8">

SimpleTest.waitForExplicitFinish();

let tests = [() => {
  /*
  {
   "iss": "323d34dc-b5cf-4822-8e47-6a4515dc74db",
    "typ": "mozilla/payments/test/success",
    "request": {
        "name": "Virtual Kiwi",
        "id": "af1f960a-3f90-4e2d-a20f-d5170aee49f2",
        "postbackURL": "https://inapp-pay-test.paas.allizom.org/mozpay",
        "productData": "localTransID=14546cd1-db9b-4759-986f-2a6a295fdcc1",
        "chargebackURL": "https://inapp-pay-test.paas.allizom.org/mozpay",
        "description": "The forbidden fruit"
    }
  }
  */
  let jwt = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiIzMjNkMzRkYy1iNWNmLTQ4MjItOGU0Ny02YTQ1MTVkYzc0ZGIiLCJyZXF1ZXN0Ijp7ImRlc2NyaXB0aW9uIjoiVGhlIGZvcmJpZGRlbiBmcnVpdCIsImlkIjoiYWYxZjk2MGEtM2Y5MC00ZTJkLWEyMGYtZDUxNzBhZWU0OWYyIiwicG9zdGJhY2tVUkwiOiJodHRwczovL2luYXBwLXBheS10ZXN0LnBhYXMuYWxsaXpvbS5vcmcvbW96cGF5L3Bvc3RiYWNrIiwicHJvZHVjdERhdGEiOiJsb2NhbFRyYW5zSUQ9MTQ1NDZjZDEtZGI5Yi00NzU5LTk4NmYtMmE2YTI5NWZkY2MxIiwiY2hhcmdlYmFja1VSTCI6Imh0dHBzOi8vaW5hcHAtcGF5LXRlc3QucGFhcy5hbGxpem9tLm9yZy9tb3pwYXkvY2hhcmdlYmFjayIsIm5hbWUiOiJWaXJ0dWFsIEtpd2kifSwidHlwIjoibW96aWxsYS9wYXltZW50cy90ZXN0L3N1Y2Nlc3MifQ.8zaeYFUCwKkZWk2TFf2wEJWrmiSYQGNbpKc2ADkvL9s";
  let req = window.navigator.mozPay(jwt);
  req.onsuccess = (result) => {
    ok(true, "Expected mozPay success");
    runTest();
  };
  req.onerror = (error) => {
    ok(false, "Unexpected mozPay error " + error);
    SimpleTest.finish();
  };
}, () => {
  /*
  {
      "iss": "323d34dc-b5cf-4822-8e47-6a4515dc74db",
      "typ": "mozilla/payments/test/failure",
      "request": {
          "name": "Virtual Kiwi",
          "id": "af1f960a-3f90-4e2d-a20f-d5170aee49f2",
          "postbackURL": "https://inapp-pay-test.paas.allizom.org/mozpay",
          "productData": "localTransID=cc3c0994-33e8-4a21-aa2c-75ee44f5fe75",
          "chargebackURL": "https://inapp-pay-test.paas.allizom.org/mozpay",
          "description": "The forbidden fruit"
      }
  }
   */
  let jwt = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiIzMjNkMzRkYy1iNWNmLTQ4MjItOGU0Ny02YTQ1MTVkYzc0ZGIiLCJyZXF1ZXN0Ijp7ImRlc2NyaXB0aW9uIjoiVGhlIGZvcmJpZGRlbiBmcnVpdCIsImlkIjoiYWYxZjk2MGEtM2Y5MC00ZTJkLWEyMGYtZDUxNzBhZWU0OWYyIiwicG9zdGJhY2tVUkwiOiJodHRwczovL2luYXBwLXBheS10ZXN0LnBhYXMuYWxsaXpvbS5vcmcvbW96cGF5L3Bvc3RiYWNrIiwicHJvZHVjdERhdGEiOiJsb2NhbFRyYW5zSUQ9Y2MzYzA5OTQtMzNlOC00YTIxLWFhMmMtNzVlZTQ0ZjVmZTc1IiwiY2hhcmdlYmFja1VSTCI6Imh0dHBzOi8vaW5hcHAtcGF5LXRlc3QucGFhcy5hbGxpem9tLm9yZy9tb3pwYXkvY2hhcmdlYmFjayIsIm5hbWUiOiJWaXJ0dWFsIEtpd2kifSwidHlwIjoibW96aWxsYS9wYXltZW50cy90ZXN0L2ZhaWx1cmUifQ.1uV4-HkmwO0oDv50wi1Ma4tNpnxoFGaw5zaPj8xkcAc";
  let req = window.navigator.mozPay(jwt);
  req.onsuccess = (result) => {
    ok(false, "Unexpected mozPay success " + result);
    SimpleTest.finish();
  };
  req.onerror = (error) => {
    ok(true, "Expected mozPay error");
    runTest();
  };
}];

function runTest() {
  if (!tests.length) {
    ok(true, "Done!");
    SimpleTest.finish();
    return;
  }
  tests.shift()();
}

const uiGlue = SpecialPowers.loadChromeScript(SimpleTest.getTestFileURL('MockPaymentsUIChromeScript.js'));
SimpleTest.registerCleanupFunction(() => {
 uiGlue.sendAsyncMessage("MockPaymentsUIGlue.cleanup", {});
});
uiGlue.sendAsyncMessage("MockPaymentsUIGlue.init", {});

SpecialPowers.pushPrefEnv({
  "set": [
    ["dom.payment.skipHTTPSCheck", "true"],
    ["dom.payment.debug", "true"],
    ["dom.payment.provider.1.name", "SuccessProvider"],
    ["dom.payment.provider.1.description", ""],
    ["dom.payment.provider.1.uri",
     "https://example.com:443/tests/dom/payment/tests/mochitest/file_payprovidersuccess.html?req="],
    ["dom.payment.provider.1.type", "mozilla/payments/test/success"],
    ["dom.payment.provider.1.requestMethod", "GET"],
    ["dom.payment.provider.2.name", "FailureProvider"],
    ["dom.payment.provider.2.description", ""],
    ["dom.payment.provider.2.uri",
     "https://example.com:443/tests/dom/payment/tests/mochitest/file_payproviderfailure.html?req="],
    ["dom.payment.provider.2.type", "mozilla/payments/test/failure"],
    ["dom.payment.provider.2.requestMethod", "GET"],
  ]
  }, () => {
    runTest();
  });

</script>
</pre>
</body>
</html>
